Move the sibling-CLI resolvers to a leaf helpers module
