+++ esphome_device_builder/controllers/editor.py
@@ -20,9 +20,9 @@ from ..helpers.api import api_command
 from ..helpers.async_ import drain_tasks, run_in_executor
 from ..helpers.json import JSONDecodeError, dumps, loads
 from ..helpers.process import kill_quietly
+from ..helpers.sibling_cli import _find_esphome_cmd
 from ..helpers.subprocess import create_subprocess_exec
 from ..models.automations import MigrateConfigResponse
-from .firmware.helpers import _find_esphome_cmd
 from .migrations import render_migrations
 
 if TYPE_CHECKING:
