+++ tests/controllers/firmware/test_verify_chip_e2e.py
@@ -15,7 +15,7 @@ The chip-id check spawns ``[*_find_esptool_cmd(), '--port', <port>,
 'chip-id']`` via ``create_subprocess_exec``. The resolved command is
 either a sibling ``esptool`` script next to ``sys.executable`` or
 ``[sys.executable, '-m', 'esptool']`` as a fallback (see
-``_find_esptool_cmd`` in ``controllers.firmware.helpers``). Tests
+``_find_esptool_cmd`` in ``helpers.sibling_cli``). Tests
 substitute ``create_subprocess_exec`` module-level so each one's
 "esptool" output can be controlled while the real subsequent build
 still runs through the same wrapper (substitute returns a no-op
